+++ gatekeeper-lib.js
