+++ keymaster-lib.js
